+++ lib/graphs.dart
@@ -1,0 +1,2 @@
+export 'src/strongly_connected_components.dart'
+    show stronglyConnectedComponents;