+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # 0.1.0-dev
 
-- Initial release
+- Initial release with an implementation of `stronglyConnectedComponents`.